REF: Merge commit message checker modules
