+++ setup.cfg
@@ -25,7 +25,7 @@ install_requires =
 
 [options.entry_points]
 console_scripts =
-    check-commit-message = conventional_commits.cli:main
+    check-commit-message = conventional_commits.commit_message_checker:main
     compile-release-notes = conventional_commits.compile_release_notes:main
     check-semantic-version = conventional_commits.check_semantic_version:main
